requirements.txt
# torch: install separately via PyTorch index URL for CPU or CUDA support.
# For CUDA 12.8: pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
torch==2.10.0
torchvision==0.25.0
Pillow==12.0.0
numpy==2.3.5
matplotlib==3.10.8
opencv-python==4.13.0.92
mlflow==3.10.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
tqdm==4.67.3
scipy==1.17.1
lime==0.2.0.1
shap==0.51.0
